+++ types/transaction.go
@@ -2,7 +2,6 @@ package types
 
 import (
 	"encoding/json"
-	"fmt"
 )
 
 type Transaction struct {
@@ -33,15 +32,4 @@ type GetChunk struct {
 
 func (gc *GetChunk) Marshal() ([]byte, error) {
 	return json.Marshal(gc)
-}
-
-type Input map[string]interface{}
-
-func (i Input) ToString() (string, error) {
-	bb, err := json.Marshal(i)
-	if err != nil {
-		fmt.Println(fmt.Errorf("json marshal input err: %v", err))
-		return "", err
-	}
-	return string(bb), nil
 }
